+++ src/index.ts
@@ -1,5 +1,6 @@
 import fs from 'fs';
-import { tools } from './lib/yaml-reader';
+import { readYamlAsJson } from './lib/yaml-reader';
+import { buildToolsFromApiConfigArray } from './lib/tools-builder';
 
 const world = 'world';
 
@@ -10,6 +11,9 @@ export function hello(who: string = world): string {
 const gretting = hello('Adolfo');
 
 console.log(gretting);
+
+const apiConfigArray = readYamlAsJson("apis.yaml").apis;
+const tools = buildToolsFromApiConfigArray(apiConfigArray);
 
 console.log(tools);
